hopefully fixing this project

+++ Assets/Scripts/Helpers.cs
@@ -7,10 +7,10 @@ namespace Assets.Scripts
 {
     public static class Helpers
     {
-        static void ToPolar(double x, double y, out float r, out float q)
+        static void ToPolar(float x, float y, out float r, out float q)
         {
             r = (float)Math.Sqrt((x * x) + (y * y));
-            q = (float)Math.Atan2(y,  x);
+            q = (float)Math.Atan2(y, x);
         }
         static void ToCartesian(float r, float q, out float x, out float y)
         {
